move chart and description columns to parent comp
